# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golf caddie backend — players, rounds, shots and course data.",
        "title": "Agentic Caddie API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/pois": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pois"
                ],
                "summary": "Create a point of interest",
                "parameters": [
                    {
                        "description": "POI to create",
                        "name": "poi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createPOIRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.poiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pois/hole/{courseHoleId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pois"
                ],
                "summary": "List POIs for a course hole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course Hole ID",
                        "name": "courseHoleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.poiResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid course hole ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "pois"
                ],
                "summary": "Delete all POIs for a course hole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course Hole ID",
                        "name": "courseHoleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid course hole ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pois/hole/{courseHoleId}/tee/{tee}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pois"
                ],
                "summary": "List POIs for a course hole filtered by tee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course Hole ID",
                        "name": "courseHoleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tee name",
                        "name": "tee",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.poiResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid course hole ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pois/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pois"
                ],
                "summary": "Get a POI by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.poiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "POI not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pois"
                ],
                "summary": "Update a point of interest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "poi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updatePOIRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "pois"
                ],
                "summary": "Delete a POI",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createPOIRequest": {
            "type": "object",
            "properties": {
                "course_hole_id": {
                    "type": "integer",
                    "example": 1
                },
                "distance_end": {
                    "type": "number",
                    "example": 200
                },
                "distance_start": {
                    "type": "number",
                    "example": 180
                },
                "label": {
                    "type": "string",
                    "example": "Fairway bunker left"
                },
                "poi_type": {
                    "type": "string",
                    "example": "bunker"
                },
                "reference_point": {
                    "type": "string",
                    "example": "tee"
                },
                "side": {
                    "type": "string",
                    "example": "left"
                },
                "specific_tee": {
                    "type": "string",
                    "example": "white"
                }
            }
        },
...
        "handlers.poiResponse": {
            "type": "object",
            "properties": {
                "course_hole_id": {
                    "type": "integer"
                },
                "distance_end": {
                    "type": "number"
                },
                "distance_start": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "poi_type": {
                    "type": "string"
                },
                "reference_point": {
                    "type": "string"
                },
                "side": {
                    "type": "string"
                },
                "specific_tee": {
                    "type": "string"
                }
            }
        },
...
        "handlers.updatePOIRequest": {
            "type": "object",
            "properties": {
                "distance_end": {
                    "type": "number",
                    "example": 200
                },
                "distance_start": {
                    "type": "number",
                    "example": 180
                },
                "label": {
                    "type": "string",
                    "example": "Fairway bunker left"
                },
                "poi_type": {
                    "type": "string",
                    "example": "bunker"
                },
                "reference_point": {
                    "type": "string",
                    "example": "tee"
                },
                "side": {
                    "type": "string",
                    "example": "left"
                },
                "specific_tee": {
                    "type": "string"
                }
            }
        },
...
    }
}